+++ src/components/Input/style/index.js
@@ -16,7 +16,7 @@ export const blockCls = prefixCls + '-block';
 export const FixWrap = styled('span')`
     vertical-align: middle;
     display: table-cell;
-    /* auto fix fixwrap to min-width */
+    /* auto fix fixWrap to min-width */
     width: 1px;
 `;
 
@@ -107,24 +107,14 @@ export const InputWrap = withProps({
     className: prefixCls
 })(
     styled('span')(props => {
-        const { disabled } = props;
-
         return css`
             position: relative;
             ${inlineBlockWithVerticalMixin};
             box-sizing: border-box;
+            font-size: 12px;
             &.${blockCls} {
                 display: block;
             }
-
-            ${disabled &&
-            css`
-                &,
-                input,
-                ${SearchIcon} {
-                    cursor: not-allowed;
-                }
-            `};
 
             input {
                 vertical-align: middle;