+++ src/components/Switch/style/index.js
@@ -195,7 +195,7 @@ const propsMixin = props => {
         ${disabled &&
         css`
             border-color: ${DT.T_COLOR_LINE_DISABLED_DARK};
-            cursor: not-allowed;
+            cursor: default;
 
             ${Inner} {
                 color: ${DT.T_COLOR_TEXT_DISABLED};
@@ -217,6 +217,7 @@ export const SwitchWrap = withProps({})(styled('div')`
     box-sizing: border-box;
     cursor: pointer;
     transition: all 0.3s;
+    font-size: 12px;
 
     ${inlineBlockWithVerticalMixin};
     ${propsMixin};